test: Add unit tests for product and user models

+++ migrations/sqls/20220802172334-products-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products_table (
     id SERIAL PRIMARY KEY,
     name VARCHAR(150),
-    price MONEY,
+    price REAL,
     category VARCHAR(100)
 );
